enable solhint for /shared/ folder
add others to ignore file
add solhint to CICD

+++ contracts/src/v0.8/shared/token/ERC677/ERC677.sol
@@ -13,7 +13,7 @@ contract ERC677 is IERC677, ERC20 {
   constructor(string memory name, string memory symbol) ERC20(name, symbol) {}
 
   /// @inheritdoc IERC677
-  function transferAndCall(address to, uint amount, bytes memory data) public returns (bool success) {
+  function transferAndCall(address to, uint256 amount, bytes memory data) public returns (bool success) {
     super.transfer(to, amount);
     emit Transfer(msg.sender, to, amount, data);
     if (to.isContract()) {
